requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
django-cleanup==8.1.0
django-htmx==1.18.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
